syntax fix

+++ src/test/java/uk/ac/vfb/geppetto/test/MultipleQueriesVFBQueryTest.java
@@ -190,12 +190,12 @@ public class MultipleQueriesVFBQueryTest
 		EList<RunnableQuery> runnableQueriesEMF = new BasicEList<RunnableQuery>();
 
 		RunnableQuery rqEMF1 = DatasourcesFactory.eINSTANCE.createRunnableQuery();
-		rqEMF1.setQueryPath(model.getQueries().get(avQ.get("partsof")).getPath());
+		rqEMF1.setQueryPath(model.getQueries().get((int) avQ.get("partsof")).getPath());
 		rqEMF1.setTargetVariablePath(variable1.getPath());
 		runnableQueriesEMF.add(rqEMF1);
 		
 		RunnableQuery rqEMF2 = DatasourcesFactory.eINSTANCE.createRunnableQuery();
-		rqEMF2.setQueryPath(model.getQueries().get(avQ.get("partsof")).getPath());
+		rqEMF2.setQueryPath(model.getQueries().get((int) avQ.get("partsof")).getPath());
 		rqEMF2.setTargetVariablePath(variable2.getPath());
 		runnableQueriesEMF.add(rqEMF2);
 
@@ -253,12 +253,12 @@ public class MultipleQueriesVFBQueryTest
 		EList<RunnableQuery> runnableQueriesEMF = new BasicEList<RunnableQuery>();
 
 		RunnableQuery rqEMF1 = DatasourcesFactory.eINSTANCE.createRunnableQuery();
-		rqEMF1.setQueryPath(model.getQueries().get(avQ.get("partsof")).getPath());
+		rqEMF1.setQueryPath(model.getQueries().get((int) avQ.get("partsof")).getPath());
 		rqEMF1.setTargetVariablePath(variable1.getPath());
 		runnableQueriesEMF.add(rqEMF1);
 		
 		RunnableQuery rqEMF2 = DatasourcesFactory.eINSTANCE.createRunnableQuery();
-		rqEMF2.setQueryPath(model.getQueries().get(avQ.get("partsof")).getPath());
+		rqEMF2.setQueryPath(model.getQueries().get((int) avQ.get("partsof")).getPath());
 		rqEMF2.setTargetVariablePath(variable2.getPath());
 		rqEMF2.setBooleanOperator(BooleanOperator.NAND);
 		runnableQueriesEMF.add(rqEMF2);
